pyproject.toml
[project]
name = "travel-planner-vercel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.59.0",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.3",
    "ddgs>=5.0.0",
    "tenacity>=8.2.3",
    "python-dotenv>=1.0.0",
]
